refactor(esbuild): wrap all the logic inside a plugin
the idea is to publish this plugin so we can centralize the logic in one place and replicate it to
all the repositories easily
